#30 WIP extracted the IcecastMetadataUpdater and merged some of the icecast configuration method into  a new superclass.
