+++ src/audio/broadcast/icecast/IcecastTCPConfigurationEditor.java
@@ -99,7 +99,7 @@ public class IcecastTCPConfigurationEditor extends BroadcastConfigurationEditor
         add(new JLabel("Mount:"));
         mMountPoint = new JTextField();
         mMountPoint.getDocument().addDocumentListener(this);
-        mMountPoint.setToolTipText("Mount point (e.g. \\12345)");
+        mMountPoint.setToolTipText("Mount point (e.g. /12345)");
         add(mMountPoint);
 
         add(new JLabel("Password:"));
